BuddyPress: add members directory field tabs
Based on the Enrolled and Association fields, two directory tabs are added
to list only members that are enrolled or are associated with the current
user. This should provide a way to interact with people only relevant to
the event: those that will attend, and those that are (somewhat) related.

Introduces a variety of field data getters, but most importantly this
commit features an effective way to couple directory tab scopes with the
associated members query. This is done through late hooking in the parsed
arguments of `has_members` and early filtering of the arguments to-parse
of `core_get_users`. It works great, for now.