BuddyPress: block restricted member profiles
For members, there shouldn't be too much information on each member
available. So to restrict this a little, the following holds. A member
profile can be viewed when:

1. The member is enrolled
2. The member is in the same association as the logged-in user

This introduces the `paco2017_bp_xprofile_association_matches()` check
which compares the association's value of two members (the displayed user
and logged-in user respectively).