Introduce Agenda logic
Implements the Agenda Item post type, the Conference Day taxonomy, and the
Conference Location taxonomy. The latter is also attached to the Workshop
post type.

Front-end presentation of post types and taxonomies is not yet supported.

Agenda Items have next to a Day and a Location, also a Time Start and Time
End post meta.