Introduce workshop users and user count
Introduces `paco2017_get_workshop_enrolled_users()` and
`paco2017_get_workshop_enrolled_user_count()` to return details on the
enrolled users that have registered for the workshop.

Adds a BuddyPress xprofile implementation that effectively combines the
data from the enrollement field and workshop fields. This way only
enrolled users are returned, while non-enrolled registrees for the
workshops are ignored.

Also adds the user count information in the workshop info box as well as
in the REST api workshop data.