BuddyPress: enrollment field types support
To determine the member's enrollment status, we previously only supported
a single-checkbox field solution. This commit fixes that.

In the Enrollment Field setting you can know define what the value for
'enrolled' is. This allows for careless field type selection and adjusting
the defined enrolled-status value for it.

Member queries for enrolled members are adjusted to work with the new and
previous solution.