Refresh enrolled users query logic
In order to enable enrolled user counts for non-loggedin users, the
statistics should exist outside of BuddyPress, which is not loaded for
them. Hense, this commit introduces a 'cache' by way of storing the
current enrolled user count in a site option.

The new `paco2017_get_enrolled_users_by_association()` is now used for
this context in the enrollments widget.